# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Risk & Disputes API
  version: 0.2.0
  description: Risk engine, fraud signals, dispute management, evidence, and representment
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3003
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/risk/rulesets/{id}:
    get:
      summary: Get risk ruleset by ID
      description: Returns a single risk ruleset.
      x-platform:
        kind: resource
        entity: RiskRuleset
        affectsLedger: false
      operationId: getRiskRulesetById
      tags:
        - RiskRules
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Risk ruleset found.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Risk ruleset not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update risk ruleset
      description: Updates an existing risk ruleset.
      x-platform:
        kind: transition
        entity: RiskRuleset
        emitsEvents:
          - risk_ruleset.updated
        affectsLedger: false
        stateChanges:
          - metadata_updated
      operationId: updateRiskRuleset
      tags:
        - RiskRules
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Risk ruleset updated.
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Risk ruleset not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete risk ruleset
      description: Deletes a risk ruleset.
      x-platform:
        kind: transition
        entity: RiskRuleset
        emitsEvents:
          - risk_ruleset.deleted
        affectsLedger: false
        stateChanges:
          - active -> deleted
      operationId: deleteRiskRuleset
      tags:
        - RiskRules
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Risk ruleset deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Risk ruleset not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/risk/lists:
    get:
      summary: List risk lists
      description: Returns a paginated list of risk lists (blocklists/allowlists).
      x-platform:
        kind: resource
        entity: RiskList
        affectsLedger: false
      operationId: listRiskLists
      tags:
        - RiskRules
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of risk lists.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a risk list
      description: Creates a new risk list (blocklist or allowlist).
      x-platform:
        kind: resource
        entity: RiskList
        emitsEvents:
          - risk_list.created
        affectsLedger: false
      operationId: createRiskList
      tags:
        - RiskRules
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum:
                    - blocklist
                    - allowlist
                description:
                  type: string
      responses:
        "201":
          description: Risk list created.
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/risk/lists/{id}:
    get:
      summary: Get risk list by ID
      description: Returns a single risk list.
      x-platform:
        kind: resource
        entity: RiskList
        affectsLedger: false
      operationId: getRiskListById
      tags:
        - RiskRules
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Risk list found.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Risk list not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update risk list
      description: Updates an existing risk list.
      x-platform:
        kind: transition
        entity: RiskList
        emitsEvents:
          - risk_list.updated
        affectsLedger: false
        stateChanges:
          - metadata_updated
      operationId: updateRiskList
      tags:
        - RiskRules
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Risk list updated.
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Risk list not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete risk list
      description: Deletes a risk list.
      x-platform:
        kind: transition
        entity: RiskList
        emitsEvents:
          - risk_list.deleted
        affectsLedger: false
        stateChanges:
          - active -> deleted
      operationId: deleteRiskList
      tags:
        - RiskRules
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Risk list deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Risk list not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...